# Change: Improve performance of finding free pool slots.