@ldpl 2024-02-11
Have you tried simply doing list of free pool slots? With OpenTTD pools never really going down much I'd expect that to be more efficient solution. Maybe a sorted list though.