@JGRennison 2024-02-11
Allocation of free pool IDs needs to be the same for all instances when in a multiplayer, maintaining a last freed structure complicates that.
Pool IDs can also be allocated directly, which also somewhat complicates maintenance of a such a structure.