@ldpl 2024-02-11
You can always take the min index from the list to avoid having to sync it in mp. And pre-fill the list up to max index on afterload.