@PeterN 2024-02-11
Here's some stats from testing with priority_queue. While it's faster in extreme cases, the performance of bitmap is sometimes faster. Wentbourne doesn't have effect vehicles due to maglev, so the primary benefit is CargoPackets.

| Action                       | Vanilla  | Bitmap | Priority_Queue |
| ---------------------------- | -------- | ------ | -------------- |
| Wentbourne CargoPacket GetNew x 10,000   | 6608us   | 38us   | 108us          |
| Wentbourne CargoPacket FreeItem x 10,000 | 7.6us    | 1.75us | 8us            |
| Xarick Vehicle GetNew x 10,000       | 113994us | 1915us | 133us          |
| Xarick Vehicle FreeItem x 10,000     | 9.3us    | 6.4us  | 13.1us         |