@LordAro 2024-02-11
Any testing in non pathological cases as well? Don't exactly want to optimise for the worst case at the expense of everything else