@PeterN 2024-02-11
Minor tweak, removed an extraneous header addition (clangd strikes again!) and moved the ClrBit() so that it's not called during pool cleaning. Not going to be noticeable but the test is already there so it's not adding anything.