@ldpl 2024-02-11
As these timings are way below TicToc precision I did some measurements with the same code I used in #7962. Afaict error is about 10% so the diffirence between bitmap and priority queue is not somethnig that can be reliably measured within the game as individual timings start to approach a single tick territory.
| Action | Vanilla |	Bitmap	| priority_queue |
| --- | -- | -- | -- |
| CargoPacket GetNew |	1986 |	238 |	266 |
| CargoPacket FreeItem |	575 |	605	| 607 |