@PeterN 2024-02-13
Performance analysis from loading a Reddit server 1 game.

![image](https://github.com/OpenTTD/OpenTTD/assets/639850/303770da-d068-4adb-adb4-abb1ca1544e7)

"It's not worse"